# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/getItems/{id}": {
            "get": {
                "description": "Get all food items from the database or a specific item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "get"
                ],
                "summary": "Get all food items or a specific item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID to get a specific item",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Model.Item"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Model.Item": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "itemBBD": {
                    "type": "integer"
                },
                "itemCount": {
                    "type": "integer"
                },
                "itemEAN": {
                    "type": "integer"
                },
                "itemName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}